% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_factor.R
\name{f_factor}
\alias{f_factor}
\title{Mandalar: package for building mandalas from parametric equations of classical curves}
\usage{
f_factor(x, y, k)
}
\arguments{
\item{x}{is a vector length n with coordinate x of point}

\item{y}{is a vector length n with coordinate y of point}

\item{k}{is a vector with factor of decrease or increase points}
}
\value{
Returns a dataframe with the original points plus the respective changed points.
}
\description{
Function to reduce points
}
\examples{
x=c(1,1)
y=c(0,1)
k=c(0.5)
f_factor(x,y,k)


}
