% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rotacao.R
\name{f_rotacao}
\alias{f_rotacao}
\title{Mandalar: package for building mandalas from parametric equations of classical curves}
\usage{
f_rotacao(x, y, rotacao)
}
\arguments{
\item{x}{is a vector length n with coordinate x of point}

\item{y}{is a vector length n with coordinate y of point}

\item{rotacao}{is a vector of length k with angles in radians to rotate the point (x,y)}
}
\value{
Returns a dataframe with the original points plus the respective rotations of these points.
}
\description{
Function to rotate points by one or more angles
}
\details{
If x and y dimension is n and rotation dimension is k, then function f_rotacao will return a dataframe with two columns and (n+1)k rows
}
\examples{
x=c(1,1)
y=c(0,1)
rotacao=c(pi/3, pi/2, pi)
f_rotacao(x,y,rotacao)

}
