% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_transxy.R
\name{f_transxy}
\alias{f_transxy}
\title{creates a dataframe containing the points for the espiral hiperbolica mandala}
\usage{
f_transxy(x, y, tx, ty)
}
\arguments{
\item{x}{is a vector length n with coordinate x of point}

\item{y}{is a vector length n with coordinate y of point}

\item{tx}{is a vector with with shifts on the x-axis}

\item{ty}{is a vector with with shifts on the y-axis}
}
\value{
Returns a dataframe with the original points plus the respective translation of these points.
}
\description{
Function to translation points by shifts on the x-axis or y-axis or both
}
\examples{
x=c(1,1)
y=c(0,1)
tx=c(-1,-2)
ty=c(0,0)
f_transxy(x,y,tx,ty)


}
\author{
Luciane Ferreira Alcoforado
}
