% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcircle.R
\name{pcircle}
\alias{pcircle}
\title{creates a dataframe containing the points for the circle mandala}
\usage{
pcircle(theta, raio, k, n)
}
\arguments{
\item{theta}{is a vector length 2 with start angle and end angle}

\item{raio}{is a vector length 1 with radius value}

\item{k}{is a vector of length 1 with angles in degree to rotate the point (x,y)}

\item{n}{is a number of points}
}
\value{
Returns a dataframe with the original points plus the respective rotations of these points.
}
\description{
Function to build point for the circle base
}
\examples{
theta = c(0,2*pi) #half turn angle
raio = 1
k = 45
n=20
pcircle(theta, raio, k, n)

}
\author{
Luciane Ferreira Alcoforado
}
