% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelipse.R
\name{pelipse}
\alias{pelipse}
\title{creates a dataframe containing the points for the elipse mandala}
\usage{
pelipse(theta, a, b, k, n)
}
\arguments{
\item{theta}{is a vector length 2 with start angle and end angle}

\item{a}{is one of the parameters of the curves; for the ellipse is the radius on the x axis}

\item{b}{is one of the parameters of the curves; for the ellipse is the radius on the y axis}

\item{k}{is a vector of length 1 with angles in degree to rotate the point (x,y)}

\item{n}{is a number of points}
}
\value{
Returns a dataframe with the original points plus the respective rotations of these points.
}
\description{
Function to build point for the elipse base
}
\examples{
theta = c(0,2*pi) #half turn angle
a = 1
b=2
k = 90
n=20
pelipse(theta, a, b, k, n)

}
\author{
Luciane Ferreira Alcoforado
}
