% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pespiral2.R
\name{pespiral2}
\alias{pespiral2}
\title{creates a dataframe containing the points for the espiral hiperbolica mandala}
\usage{
pespiral2(theta, raio, k, n)
}
\arguments{
\item{theta}{is a vector length 2 with start angle and end angle}

\item{raio}{is a vector length 1 with radius value}

\item{k}{is a vector of length 1 with angles in degree to rotate the point (x,y)}

\item{n}{is a number of points}
}
\value{
Returns a dataframe with the original points plus the respective rotations of these points.
}
\description{
Function to build a espiral hiperbolica
}
\examples{
theta = c(0,6*pi)
k=45
raio = 1
n=20
pespiral2(theta, raio, k, n)

}
\author{
Luciane Ferreira Alcoforado
}
