% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plemniscata.R
\name{plemniscata}
\alias{plemniscata}
\title{creates a dataframe containing the points for the lemniscata mandala}
\usage{
plemniscata(theta, raio, k, n)
}
\arguments{
\item{theta}{is a vector length 2 with start angle and end angle}

\item{raio}{is a vector length 1 with radius value}

\item{k}{is a vector of length 1 with angles in degree to rotate the point (x,y)}

\item{n}{is a number of points}
}
\value{
Returns a dataframe with the original points plus the respective rotations of these points.
}
\description{
Function to build a lemniscata
}
\examples{
theta = c(0,2*pi)
k=45
raio = 1
n=20
plemniscata(theta, raio, k, n)

}
\author{
Luciane Ferreira Alcoforado
}
