% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mandala.R
\name{plot_mandala}
\alias{plot_mandala}
\title{creates a mandala visualization}
\arguments{
\item{dt}{dataframe with points x and y}
}
\value{
Returns a plot
}
\description{
Function to plot a mandala with points in dataframe
}
\examples{
require(ggplot2)
n=500; raio=1; t=seq(0,2*pi, length.out = n)
x1=raio*cos(t)
y1=raio*sin(t)
#pontos para os 3 círculos: translação dos pontos iniciais (x1,x=c(x1,x1-raio,x1-2*raio)
x=c(x1,x1-raio,x1-2*raio)
y=c(y1,y1,y1)
dt=data.frame(x,y,z="circulo")
rotacao = (pi/8)*(1:16); n=length(x); xt1=x; yt1=y
dt=f_rotacao(x=dt$x, y=dt$y, rotacao)
plot_mandala(dt)
}
\author{
Luciane Ferreira Alcoforado
}
